Created folder for imageGallery components and moved styled to bottom of files
