+++ client/src/components/overview/imageGallery/ExpandedView.jsx
@@ -1,37 +1,9 @@
 import React, { useState } from 'react';
-import styled from 'styled-components';
 import { PropTypes } from 'prop-types';
-import { StyledLeftArrowExpand, StyledRightArrowExpand, StyledUpArrowExpand, StyledDownArrowExpand, StyledBackButton } from './styledIcons.js';
+import styled from 'styled-components';
+import { StyledLeftArrowExpand, StyledRightArrowExpand, StyledUpArrowExpand, StyledDownArrowExpand, StyledBackButton } from '../styledIcons.js';
 import ImageCarousel from './ImageCarousel.jsx';
 import ImageSidebar from './ImageSidebar.jsx';
-
-const StyledCarousel = styled.section`
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`
-
-const StyledThumbnailAlign = styled.div`
-  position: relative;
-`
-
-const StyledCarouselImageSize = styled.div`
-  width: 500px;
-  &:hover {
-    cursor:crosshair;
-  }
-`
-
-const StyledButtonDiv = styled.div`
-  display: flex;
-  justify-content: center;
-`
-
-const StyledHorizontalButtons = styled.div`
-  display: flex;
-  align-content: center;
-`
 
 const ExpandedView = ( {styleImages, activeThumbnails, current, setCurrent, nextSlide, prevSlide, verticalScroll, upSlide, downSlide, length, start, end, changeView} ) => {
   const [magnified, setMagnified] = useState(false);
@@ -81,5 +53,33 @@ ExpandedView.propTypes = {
   end: PropTypes.number,
   changeView: PropTypes.func
 }
+
+const StyledCarousel = styled.section`
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`
+
+const StyledThumbnailAlign = styled.div`
+  position: relative;
+`
+
+const StyledCarouselImageSize = styled.div`
+  width: 500px;
+  &:hover {
+    cursor:crosshair;
+  }
+`
+
+const StyledButtonDiv = styled.div`
+  display: flex;
+  justify-content: center;
+`
+
+const StyledHorizontalButtons = styled.div`
+  display: flex;
+  align-content: center;
+`
 
 export default ExpandedView;
